# travis, build: speed up CI runs