Work isn't fun if I have to wait 50 minutes for a test run. This PR is supposed to change that.
I think 10 minute test times are achievable. To get there we need to skip non-essential jobs for PRs
and speed up some package tests.

The slowest packages are:

   - [ ] eth/downloader (216.364s)
   - [ ] cmd/swarm (194.770s)
   - [ ] tests (138.679s)
   - [ ] rpc (63.584s)
   - [ ] swarm/storage (59.564s)
   - [ ] swarm (55.446s)
   - [x] les (54.602s)
   - [ ] whisper/whisperv5 (46.572s)
   - [ ] swarm/network/stream (45.232s)